moved and renamed clean script
